+++ Server/TaxasServer/src/TaxasRoomState.cpp
@@ -6,220 +6,6 @@
 #include "TaxasServerApp.h"
 #include <time.h>
 #include "TaxasPlayer.h"
-//CTaxasBaseRoomState::CTaxasBaseRoomState()
-//{
-//	m_fDuringTime = 0 ;
-//	m_pRoom = NULL ;
-//}
-//
-//CTaxasBaseRoomState::~CTaxasBaseRoomState()
-//{
-//
-//}
-//
-//bool CTaxasBaseRoomState::OnMessage( stMsg* prealMsg , eMsgPort eSenderPort , uint32_t nPlayerSessionID )
-//{
-//	switch( prealMsg->usMsgType )
-//	{
-//	case MSG_TP_PLAYER_SIT_DOWN:
-//		{
-//			stMsgTaxasPlayerSitDown* pRet = (stMsgTaxasPlayerSitDown*)prealMsg ;
-//			stMsgTaxasPlayerSitDownRet msgBack ;
-//			msgBack.nRet = 0 ;
-//			msgBack.nSeatIdx = pRet->nSeatIdx ;
-//			if ( m_pRoom->IsPlayerInRoomWithSessionID(nPlayerSessionID) == false )
-//			{
-//				msgBack.nRet = 2;
-//				m_pRoom->SendMsgToPlayer(nPlayerSessionID,&msgBack,sizeof(msgBack)) ;
-//				CLogMgr::SharedLogMgr()->ErrorLog("you are not in room session id = %d",nPlayerSessionID ) ;
-//				return true ;
-//			}
-//
-//			if ( m_pRoom->isPlayerAlreadySitDown(nPlayerSessionID) )
-//			{
-//				msgBack.nRet = 4;
-//				m_pRoom->SendMsgToPlayer(nPlayerSessionID,&msgBack,sizeof(msgBack)) ;
-//				CLogMgr::SharedLogMgr()->ErrorLog("you are already sit down session id = %d",nPlayerSessionID ) ;
-//				return true ;
-//			}
-//
-//			if ( pRet->nSeatIdx >= m_pRoom->m_stRoomConfig.nMaxSeat || m_pRoom->m_vSitDownPlayers[pRet->nSeatIdx].IsInvalid() == false )
-//			{
-//				msgBack.nRet = 1;
-//				m_pRoom->SendMsgToPlayer(nPlayerSessionID,&msgBack,sizeof(msgBack)) ;
-//				CLogMgr::SharedLogMgr()->PrintLog("invalid seat idx = %d , can not sit down session id = %d",pRet->nSeatIdx,nPlayerSessionID ) ;
-//				return true ;
-//			}
-//
-//			if ( pRet->nTakeInMoney > m_pRoom->m_stRoomConfig.nMaxTakeInCoin || pRet->nTakeInMoney < m_pRoom->m_stRoomConfig.nMiniTakeInCoin )
-//			{
-//				msgBack.nRet = 3;
-//				m_pRoom->SendMsgToPlayer(nPlayerSessionID,&msgBack,sizeof(msgBack)) ;
-//				CLogMgr::SharedLogMgr()->PrintLog("invalid take in coin = %I64d , can not sit down session id = %d",pRet->nTakeInMoney,nPlayerSessionID ) ;
-//				return true ;
-//			}
-//
-//			stTaxasInRoomPeerDataExten* pData = m_pRoom->GetInRoomPlayerDataBySessionID(nPlayerSessionID);
-//			if ( pData == nullptr )
-//			{
-//				msgBack.nRet = 2;
-//				m_pRoom->SendMsgToPlayer(nPlayerSessionID,&msgBack,sizeof(msgBack)) ;
-//				CLogMgr::SharedLogMgr()->ErrorLog("you are not in room session id = %d data is null",nPlayerSessionID ) ;
-//				return true ;
-//			}
-//
-//			m_pRoom->OnPlayerSitDown(pRet->nSeatIdx,nPlayerSessionID,pRet->nTakeInMoney) ;
-//			CLogMgr::SharedLogMgr()->PrintLog("player seat idx = %d sit down want takein coin = %I64d",pRet->nSeatIdx,pRet->nTakeInMoney) ;
-//			return true ;
-//		}
-//		break;
-//	case MSG_TP_PLAYER_STAND_UP:
-//		{
-//			uint8_t nSeatIdx = m_pRoom->GetSeatIdxBySessionID(nPlayerSessionID);
-//			if ( nSeatIdx >= m_pRoom->m_stRoomConfig.nMaxSeat )
-//			{
-//				CLogMgr::SharedLogMgr()->ErrorLog("you already stand up session id = %d",nPlayerSessionID ) ;
-//				break;
-//			}
-//			m_pRoom->OnPlayerStandUp(nSeatIdx) ;
-//			CLogMgr::SharedLogMgr()->PrintLog("player seat idx = %d stand up ",nSeatIdx );
-//		}
-//		break;
-//	case MSG_TP_ORDER_LEAVE:
-//	case MSG_TP_PLAYER_LEAVE:
-//		{
-//			if ( m_pRoom->IsPlayerInRoomWithSessionID(nPlayerSessionID) == false )
-//			{
-//				// already check by top level code ;
-//				CLogMgr::SharedLogMgr()->ErrorLog("you are not in room so how you leave , session id = %d",nPlayerSessionID ) ;
-//				break;
-//			}
-//			m_pRoom->OnPlayerLeaveRoom(nPlayerSessionID) ;
-//			CLogMgr::SharedLogMgr()->PrintLog("player session id = %d leave room ",nPlayerSessionID ) ;
-//		}
-//		break;
-//	default:
-//		return false ;
-//	}
-//	return true ;
-//}
-//
-//void CTaxasBaseRoomState::EnterState(CTaxasRoom* pRoom )
-//{
-//	m_pRoom = pRoom ;
-//}
-//
-//void CTaxasBaseRoomState::LeaveState()
-//{
-//	m_fDuringTime = 0 ;
-//}
-//
-//void CTaxasBaseRoomState::Update(float fDelte )
-//{
-//	m_fDuringTime -= fDelte ;
-//	if ( m_fDuringTime <= 0 )
-//	{
-//		OnStateTimeOut();
-//	}
-//}
-//
-//void CTaxasStateDead::EnterState(CTaxasRoom* pRoom )
-//{
-//	CTaxasBaseRoomState::EnterState(pRoom);
-//	// send inform ;
-//	if ( pRoom->getOwnerUID() == MATCH_MGR_UID )  // match room 
-//	{
-//		pRoom->onMatchFinish();
-//		m_MatchRoomDuringTime = pRoom->getDeadTime() - pRoom->getCreateTime();
-//		m_fMatchRestarTime = TIME_MATCH_PAUSE ; // 30 min later start new match ;
-//	}
-//	else
-//	{
-//		pRoom->sendExpireInform();
-//	}
-//}
-//
-//void CTaxasStateDead::Update(float fDelte )
-//{
-//	if ( m_pRoom->isDeleteRoom() == false && m_pRoom->getOwnerUID() != MATCH_MGR_UID )
-//	{
-//		time_t deadTime = m_pRoom->getDeadTime() ;
-//		if ( time(nullptr) - deadTime > 8640*3 ) // 3 days ;
-//		{
-//			m_pRoom->deleteRoom();
-//		}
-//	}
-//
-//	if ( m_pRoom->isRoomAlive() )
-//	{
-//		m_pRoom->GoToState(eRoomState_TP_WaitJoin) ;
-//		return ;
-//	}
-//	else
-//	{
-//		if ( m_pRoom->getOwnerUID() == MATCH_MGR_UID )
-//		{
-//			m_fMatchRestarTime -= fDelte;
-//			if ( m_fMatchRestarTime <= 0 )
-//			{
-//				m_pRoom->setCreateTime(time(nullptr));
-//				m_pRoom->setDeadTime(m_pRoom->getCreateTime() + m_MatchRoomDuringTime) ;
-//				m_pRoom->onMatchRestart();
-//			}
-//		}
-//	}
-//}
-//
-//// wait join state 
-//void CTaxasStateWaitJoin::EnterState(CTaxasRoom* pRoom )
-//{
-//	CTaxasBaseRoomState::EnterState(pRoom);
-//	m_pRoom->ResetRoomData() ;
-//	m_pRoom->debugPlayerHistory();
-//	CLogMgr::SharedLogMgr()->PrintLog("CTaxasStateWaitJoin");
-//}
-//
-//void CTaxasStateWaitJoin::Update(float fDelte )
-//{
-//	if ( !m_pRoom->isRoomAlive() )
-//	{
-//		m_pRoom->GoToState(eRoomState_TP_Dead) ;
-//		return ;
-//	}
-//
-//	if ( m_pRoom->GetPlayerCntWithState(eRoomPeer_WaitNextGame) >= 2 )
-//	{
-//		m_pRoom->GoToState(eRoomState_TP_BetBlind) ;
-//	}
-//}
-
-// start blind bet state 
-//void CTaxasStateBlindBet::EnterState(CTaxasRoom* pRoom )
-//{
-//	CTaxasBaseRoomState::EnterState(pRoom);
-//	m_fDuringTime = TIME_BLIND_BET_STATE + 1.5f;
-//	pRoom->StartGame();
-//	CLogMgr::SharedLogMgr()->PrintLog("CTaxasStateBlindBet");
-//}
-//
-//void CTaxasStateBlindBet::OnStateTimeOut()
-//{
-//	if ( m_pRoom->GetPlayerCntWithState(eRoomPeer_WaitCaculate) == 0 )
-//	{
-//		m_pRoom->GoToState(eRoomState_TP_WaitJoin) ;
-//		return ;
-//	}
-//
-//	if ( m_pRoom->GetPlayerCntWithState(eRoomPeer_CanAct) >= 2 )
-//	{
-//		m_pRoom->GoToState(eRoomState_TP_PrivateCard) ;
-//	}
-//	else
-//	{
-//		m_pRoom->GoToState(eRoomState_TP_OneRoundBetEndResult) ;
-//	}
-//}
-
 // start game private card
 void CTaxasStateStartGame::enterState(IRoom* pRoom )
 {
@@ -275,7 +61,7 @@ void CTaxasStatePlayerBet::update(float fDelte )
 			m_bIsDoFinished = true ;
 			if ( m_pRoom->getMostBetCoinThisRound() > 0 )
 			{
-				setStateDuringTime(TIME_PLAYER_BET_COIN_ANI + 0.2);
+				setStateDuringTime(TIME_PLAYER_BET_COIN_ANI + 0.2f);
 			}
 			else
 			{
@@ -303,7 +89,7 @@ void CTaxasStatePlayerBet::onStateDuringTimeUp()
 		return ;
 	}
 
-	uint8_t nWaitCaPlyCnt = m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
+	uint8_t nWaitCaPlyCnt = (uint8_t)m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
 	if ( nWaitCaPlyCnt == 0 )
 	{
 		CLogMgr::SharedLogMgr()->ErrorLog("why all player leave in one minite room id = %d",m_pRoom->getRoomID()) ;
@@ -361,7 +147,7 @@ bool CTaxasStatePlayerBet::onMessage( stMsg* prealMsg , eMsgPort eSenderPort , u
 			{
 				CLogMgr::SharedLogMgr()->SystemLog("client should send give up msg , before player standup or leave room  player uid = %d",pPlayer->getUserUID()) ;
 				// do something ;
-				uint64_t nValue = 0 ;
+				uint32_t nValue = 0 ;
 				m_pRoom->OnPlayerAction(pPlayer->getIdx(),eRoomPeerAction_GiveUp,nValue);
 				playerDoActOver() ;
 				return false ;
@@ -399,7 +185,7 @@ bool CTaxasStatePlayerBet::onMessage( stMsg* prealMsg , eMsgPort eSenderPort , u
 			if ( msgBack.nRet == 0 )
 			{
 				playerDoActOver();
-				CLogMgr::SharedLogMgr()->PrintLog("room id = %d player idx = %d do act = %d, value = %I64d",m_pRoom->getRoomID(), nSeatIdx,pAct->nPlayerAct,pAct->nValue ) ;
+				CLogMgr::SharedLogMgr()->PrintLog("room id = %d player idx = %d do act = %d, value = %u",m_pRoom->getRoomID(), nSeatIdx,pAct->nPlayerAct,pAct->nValue ) ;
 			}
 			else
 			{
@@ -426,7 +212,7 @@ void CTaxasStatePlayerBet::playerDoActOver()
 		m_bIsDoFinished = true ;
 		if ( m_pRoom->getMostBetCoinThisRound() > 0 )
 		{
-			setStateDuringTime(TIME_PLAYER_BET_COIN_ANI + 0.2);
+			setStateDuringTime(TIME_PLAYER_BET_COIN_ANI + 0.2f);
 		}
 		else
 		{
@@ -448,7 +234,7 @@ void CTaxasStateOneRoundBetEndResult::enterState(IRoom* pRoom)
 
 void CTaxasStateOneRoundBetEndResult::onStateDuringTimeUp()
 {
-	uint8_t nWaitCaPlyCnt = m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
+	uint8_t nWaitCaPlyCnt = (uint8_t)m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
 	if ( nWaitCaPlyCnt == 0 )
 	{
 		CLogMgr::SharedLogMgr()->ErrorLog("why all player leave in one minite room id = %d",m_pRoom->getRoomID()) ;
@@ -491,7 +277,7 @@ void CTaxasStatePublicCard::onStateDuringTimeUp()
 		return ;
 	}
 
-	uint8_t nWaitCaPlyCnt = m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
+	uint8_t nWaitCaPlyCnt = (uint8_t)m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
 	if ( nWaitCaPlyCnt == 0 )
 	{
 		CLogMgr::SharedLogMgr()->ErrorLog("why all player leave in one minite room id = %d",m_pRoom->getRoomID()) ;
@@ -514,7 +300,7 @@ void CTaxasStateGameResult::enterState(IRoom* pRoom)
 {
 	CLogMgr::SharedLogMgr()->PrintLog("room id = %d enter CTaxasStateGameResult",pRoom->getRoomID() );
 	m_pRoom = (CTaxasRoom*)pRoom ;
-    uint8_t nWaitCal = m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
+    uint8_t nWaitCal = (uint8_t)m_pRoom->getPlayerCntWithState(eRoomPeer_WaitCaculate);
 	float fTime = 0 ;
 	if ( nWaitCal > 1 )
 	{
